Validando Eliminar Genero

+++ src/main/java/com/metabit/planilla/controller/GeneroController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.metabit.planilla.entity.Genero;
+import com.metabit.planilla.service.EmpleadoService;
 import com.metabit.planilla.service.GeneroService;
 
 @Controller
@@ -19,12 +20,17 @@ public class GeneroController {
 	
 	@Autowired
 	@Qualifier("generoServiceImpl")
-	private GeneroService generoService;
+	private GeneroService generoService; 
 
+	@Autowired
+	@Qualifier("empleadoServiceImpl")
+	private EmpleadoService empleadoService;
+	
 	@GetMapping("/index")
 	private String index(Model model, @RequestParam(name="store_success", required=false) String store_success, 
 			@RequestParam(name="update_success", required=false) String update_success,
-			@RequestParam(name="delete_success", required=false) String delete_success) {
+			@RequestParam(name="delete_success", required=false) String delete_success,
+			@RequestParam(name="delete_restricted", required=false) String delete_restricted) {
 		
 		//Recibiendo el posible param de Exito en actualizacion
 		model.addAttribute("update_success", update_success);
@@ -34,6 +40,9 @@ public class GeneroController {
 		
 		//Recibiendo el posible param de Exito en el eliminacion
 		model.addAttribute("delete_success", delete_success);
+		
+		//Recibiendo el posible param de Restriccion en la eliminacion
+		model.addAttribute("delete_restricted", delete_restricted);
 		
 		//Mandando el objeto de Genero, el cual se utilizara en caso de insertar uno nuevo con el formulario respectivo
 		model.addAttribute("generoEntity", new Genero());
@@ -72,6 +81,15 @@ public class GeneroController {
 	
 	@PostMapping("/destroy")
 	private String destroy(@RequestParam("idGeneroDestroy") int idGenero) {
+		//Validacion, si hay empleados con este genero, no se debe eliminar.
+		Genero genero_selected = generoService.getGenero(idGenero);
+		
+		//Obtenemos la cantidad de empleados con este genero
+		int cant_empleados = empleadoService.findByGenero(genero_selected).size();
+	
+		//Si la cantidad es mayor a cero, significa que no podemos eliminar el genero
+		if(cant_empleados > 0 )
+			return "redirect:/planilla/genero/index?delete_restricted";
 		
 		//eliminar el genero por el idGenero
 		generoService.deleteGenero(idGenero);
